Add fire piranha plant

+++ Scripts/Projectiles/Fireball.gd
@@ -5,6 +5,7 @@ var vis: VisibilityEnabler2D = VisibilityEnabler2D.new()
 var dir: int = 1
 var velocity: Vector2 = Vector2(426, 0)
 var skip_frame: bool = false
+var gravity_scale: float = 1
 
 var belongs: int = 0 # 0 - Mario, 1 - Fire Piranha Plant, 2 - Fire Bro
 
@@ -17,23 +18,33 @@ func _ready() -> void:
 func _physics_process(delta) -> void:
   var overlaps = $CollisionArea.get_overlapping_bodies()
 
-  if overlaps.size() > 0:
+  if overlaps.size() > 0 and belongs == 0:
     for i in range(overlaps.size()):
       if overlaps[i].is_in_group('Enemy') and overlaps[i].has_method('kill'):
         overlaps[i].kill(AliveObject.DEATH_TYPE.FALL, 0)
         explode()
+        
+  if belongs != 0 and is_mario_collide('InsideDetector'):
+    Global._ppd()
 
-  if is_on_floor():
+  if is_on_floor() and belongs != 1:
     velocity.y = -350
 
-  velocity.y += 24 * Global.get_delta(delta)
+  velocity.y += 24 * gravity_scale * Global.get_delta(delta)
 
-  velocity = move_and_slide(velocity, Vector2.UP)
+  if belongs != 1:
+    velocity = move_and_slide(velocity, Vector2.UP)
+  else:
+    position += velocity * Vector2(delta, delta)
 
-  if (is_on_wall() or velocity.x == 0) and belongs == 0:
+  if (is_on_wall() or velocity.x == 0) and belongs != 1:
     explode()
   
   $Sprite.rotation_degrees += 12 * (-1 if velocity.x < 0 else 1) * Global.get_delta(delta)
+  
+func is_mario_collide(_detector_name: String) -> bool:
+  var collisions = Global.Mario.get_node(_detector_name).get_overlapping_bodies()
+  return collisions && collisions.has(self)
 
 func explode() -> void:
   var explosion = Explosion.new(position)
